[FE] 상하단 메뉴바 접근성 개선 (#777)
* feat: 하단 메뉴바에 접근성을 위해 목록 추가, 레이블 추가

* feat: 상단 메뉴바에 프로필 버튼 레이블 추가

* refactor: 모바일 상단 메뉴바 컴포넌트 위치 headerNav로 이동, 레이블 추가
