+++ frontend/src/components/common/BottomNavigation/BottomNavigation.tsx
@@ -4,10 +4,18 @@ import ROUTES from '@/constants/routes';
 
 function BottomNavigation() {
   return (
-    <S.Container>
-      <S.NavButton to={ROUTES.PRODUCTS}>제품 검색</S.NavButton>
-      <S.NavButton to={ROUTES.HOME}>홈</S.NavButton>
-      <S.NavButton to={ROUTES.PROFILE_SEARCH}>프로필 검색</S.NavButton>
+    <S.Container aria-label={'하단 메뉴바'}>
+      <S.NavList>
+        <li>
+          <S.NavButton to={ROUTES.PRODUCTS}>제품 검색</S.NavButton>
+        </li>
+        <li>
+          <S.NavButton to={ROUTES.HOME}>홈</S.NavButton>
+        </li>
+        <li>
+          <S.NavButton to={ROUTES.PROFILE_SEARCH}>프로필 검색</S.NavButton>
+        </li>
+      </S.NavList>
     </S.Container>
   );
 }
